Refactor imports to use domain.datamodels for consistency across modules
