+++ libs/commands/graph/personal.py
@@ -13,7 +13,7 @@ from plotly.subplots import make_subplots  # type: ignore
 
 import libs.global_value as g
 from libs.data import loader
-from libs.datamodels import GameInfo
+from libs.domain.datamodels import GameInfo
 from libs.functions import compose, message
 from libs.types import StyleOptions
 from libs.utils import formatter, graphutil, textutil
