+++ libs/commands/graph/rating.py
@@ -9,7 +9,7 @@ import plotly.express as px  # type: ignore
 
 import libs.global_value as g
 from libs.data import aggregate, loader
-from libs.datamodels import GameInfo
+from libs.domain.datamodels import GameInfo
 from libs.functions import compose, message
 from libs.types import StyleOptions
 from libs.utils import formatter, graphutil, textutil
