+++ libs/commands/graph/summary.py
@@ -11,7 +11,7 @@ import plotly.express as px  # type: ignore
 
 import libs.global_value as g
 from libs.data import loader
-from libs.datamodels import GameInfo
+from libs.domain.datamodels import GameInfo
 from libs.functions import compose, message
 from libs.types import StyleOptions
 from libs.utils import formatter, graphutil, textutil
