+++ libs/commands/report/matrix.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 import libs.global_value as g
 from libs.data import aggregate
-from libs.datamodels import GameInfo
+from libs.domain.datamodels import GameInfo
 from libs.functions import message
 from libs.types import StyleOptions
 from libs.utils import formatter, textutil
