+++ libs/commands/report/stats_list.py
@@ -8,7 +8,7 @@ import matplotlib.pyplot as plt
 
 import libs.global_value as g
 from libs.data import loader
-from libs.datamodels import GameInfo
+from libs.domain.datamodels import GameInfo
 from libs.functions import compose, message
 from libs.types import StyleOptions
 from libs.utils import formatter, graphutil, textutil
