+++ libs/commands/results/summary.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 import libs.global_value as g
 from libs.data import aggregate, loader
-from libs.datamodels import GameInfo
+from libs.domain.datamodels import GameInfo
 from libs.functions import message
 from libs.types import StyleOptions
 from libs.utils import converter, formatter
