+++ libs/domain/__init__.py
@@ -1,5 +1,6 @@
 """アプリケーションロジック
 
+- `libs.domain.datamodels`: データモデル
 - `libs.domain.command`: コマンド引数解析
 - `libs.domain.score`: スコア解析
 - `libs.domain.stats`: 成績解析
